V2.1
Sprint 1: Correção de 5/7 dos débitos técnicos
Sprint 2: Cadastro e alteração de produtos com UMA imagem
-------------------------------
Falta página de detalhe, carrossel, login, autenticação
