--- iOutra/src/main/java/com/example/iOutra/Seguranca/SegConfig.java
@@ -1,68 +1,0 @@
-package com.example.iOutra.Seguranca;
-
-import com.example.iOutra.servico.UserDetailsService;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.AuthenticationProvider;
-import org.springframework.security.authentication.dao.DaoAuthenticationProvider;
-import org.springframework.security.config.Customizer;
-import org.springframework.security.config.annotation.authentication.configuration.AuthenticationConfiguration;
-import org.springframework.security.config.annotation.web.builders.HttpSecurity;
-import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
-import org.springframework.security.config.annotation.web.configurers.AbstractHttpConfigurer;
-import org.springframework.security.config.http.SessionCreationPolicy;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
-import org.springframework.security.crypto.password.PasswordEncoder;
-import org.springframework.security.web.SecurityFilterChain;
-import org.springframework.security.web.authentication.UsernamePasswordAuthenticationFilter;
-
-@Configuration
-@EnableWebSecurity
-public class SegConfig {
-
-    @Autowired
-    private UserDetailsService userDetailsService;
-    @Autowired
-    private JWTauthFilter jwtAuthFilter;
-
-   @Bean
-    public SecurityFilterChain securityFilterChain(HttpSecurity httpSecurity) throws Exception{
-       httpSecurity.csrf(AbstractHttpConfigurer::disable)
-               .cors(Customizer.withDefaults())
-               .authorizeHttpRequests(request-> request.requestMatchers("/auth/**", "/public/**").permitAll()
-                       .requestMatchers("/admin/**").hasAnyAuthority("ADMIN")
-                       .requestMatchers("/user/**").hasAnyAuthority("USER")
-                       .requestMatchers("/adminuser/**").hasAnyAuthority("ADMIN", "USER")
-                       .anyRequest().authenticated())
-               .sessionManagement(manager->manager.sessionCreationPolicy(SessionCreationPolicy.STATELESS))
-               .authenticationProvider(authenticationProvider()).addFilterBefore(
-                       jwtAuthFilter, UsernamePasswordAuthenticationFilter.class
-               );
-       return httpSecurity.build();
-   }
-@Bean
-    public AuthenticationProvider authenticationProvider(){
-    DaoAuthenticationProvider daoAuthenticationProvider = new DaoAuthenticationProvider();
-    daoAuthenticationProvider.setUserDetailsService(userDetailsService);
-    daoAuthenticationProvider.setPasswordEncoder(passwordEncoder());
-    return daoAuthenticationProvider;
-}
-
-@Bean
-    public PasswordEncoder passwordEncoder(){
-       return new BCryptPasswordEncoder();
-
-}
-@Bean
-    public AuthenticationManager authenticationManager(AuthenticationConfiguration authenticationConfiguration) throws Exception{
-       return authenticationConfiguration.getAuthenticationManager();
-}
-
-
-
-}
-
-
-
