--- iOutra/src/main/java/com/example/iOutra/controller/LoginController.java
@@ -1,16 +1,0 @@
-package com.example.iOutra.controller;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.autoconfigure.neo4j.Neo4jProperties;
-import org.springframework.boot.autoconfigure.pulsar.PulsarProperties;
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class LoginController {
-
-@GetMapping("/login")
-public String login(){
-    return "login";
-}
-}
