--- iOutra/src/main/java/com/example/iOutra/model/Role.java
@@ -1,7 +1,0 @@
-package com.example.iOutra.model;
-
-public enum Role {
-    ROLE_USER,
-    ROLE_ADMIN,
-    ROLE_ESTOQUISTA
-}
