--- iOutra/src/main/java/com/example/iOutra/servico/JWTUtils.java
@@ -1,65 +1,0 @@
-package com.example.iOutra.servico;
-
-import io.jsonwebtoken.Claims;
-import io.jsonwebtoken.Jwts;
-import org.springframework.security.core.userdetails.User;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.stereotype.Component;
-
-import javax.crypto.SecretKey;
-import javax.crypto.spec.SecretKeySpec;
-import java.nio.charset.StandardCharsets;
-import java.util.Base64;
-import java.util.Date;
-import java.util.HashMap;
-import java.util.function.Function;
-
-import static io.jsonwebtoken.Jwts.claims;
-
-@Component
-public class JWTUtils {
-
-    private SecretKey Chave;
-    private static final long EXPIRATION_TIME = 8640000; //24 horas
-
-    public JWTUtils(){
-        String secreteString = "843567893696976453275974432697R634976R738467TR678T34865R6834R8763T478378637664538745673865783678548735687R3";
-        byte[] keyBytes = Base64.getDecoder().decode(secreteString.getBytes(StandardCharsets.UTF_8));
-        this.Chave = new SecretKeySpec(keyBytes, "HmacSHA256");
-
-    }
-
-    public String gerarToken(UserDetails userDetails){
-        return Jwts.builder()
-                .subject(userDetails.getUsername())
-                .issuedAt(new Date(System.currentTimeMillis()))
-                .expiration(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
-                .signWith(Chave)
-                .compact();
-    }
-
-    public String gerarRefreshToken(HashMap<String, Object> claims, UserDetails userDetails){
-        return Jwts.builder()
-                .claims(claims)
-                .subject(userDetails.getUsername())
-                .issuedAt(new Date(System.currentTimeMillis()))
-                .expiration(new Date(System.currentTimeMillis() + EXPIRATION_TIME))
-                .signWith(Chave)
-                .compact();
-    }
-
-    public String extractUsername(String token){
-        return extractClaims(token, Claims::getSubject );
-    }
-    private <T> T extractClaims(String token, Function<Claims, T> claimsTFunction){
-        return claimsTFunction.apply(Jwts.parser().verifyWith(Chave).build().parseSignedClaims(token).getPayload());
-    }
-    public  boolean isTokenValid(String token, UserDetails userDetails){
-        final String username = extractUsername(token);
-        return (username.equals(userDetails.getUsername()) && !isTokenExpired(token));
-    }
-
-    public  boolean isTokenExpired(String token){
-        return extractClaims(token, Claims::getExpiration).before(new Date());
-    }
-}
