--- iOutra/src/main/java/com/example/iOutra/servico/UserDetailsService.java
@@ -1,19 +1,0 @@
-package com.example.iOutra.servico;
-
-import com.example.iOutra.repository.UsuarioRepository;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.security.core.userdetails.UserDetails;
-import org.springframework.security.core.userdetails.UsernameNotFoundException;
-import org.springframework.stereotype.Service;
-
-@Service
-public class UserDetailsService implements org.springframework.security.core.userdetails.UserDetailsService {
-
-    @Autowired
-    private UsuarioRepository usuarioRepository;
-
-    @Override
-    public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
-        return usuarioRepository.findByEmail(username);
-    }
-}
