--- iOutra/src/main/resources/application.properties
@@ -1,22 +1,0 @@
-spring.application.name=iOutra
-
-# Configuracoes do Banco de dados
-spring.sql.init.platform=h2
-spring.datasource.driver-class-name=org.h2.Driver
-
-spring.datasource.url=jdbc:h2:file:~/ioutra-bd
-spring.datasource.username=sa
-spring.datasource.password=
-
-spring.jpa.show-sql=true
-spring.jpa.hibernate.ddl-auto=update
-#spring.jpa.open-in-view=false
-
-# Configuracoes especificas do banco H2
-spring.h2.console.enabled=true
-spring.h2.console.settings.web-allow-others=true
-spring.h2.console.path=/h2-console
-
-# JACKSON 2 - FORMATAR JSON NA RESPOSTA
-spring.jackson.serialization.INDENT_OUTPUT=true
-
