--- iOutra/src/main/resources/templates/cadastrarUsuario.html
@@ -1,97 +1,0 @@
-<!DOCTYPE html>
-<html lang="pt-br">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>iOutra - Novo Usuário</title>
-
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
-        integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
-</head>
-
-<body>
-    <div class="container">
-        <div class="row">
-            <div class="col-md-8 mx-auto rounded border p-4 m-4">
-                <h2 class="text-center mb-5">Novo Usuário</h2>
-
-                <form method="post" th:object="${usuarioDTO}">
-
-                    <div class="row mb-3">
-                        <label class="col-sm-4 col-form-label">Nome</label>
-                        <div class="col-sm-8">
-                            <input class="form-control" th:field="${usuarioDTO.nome}">
-                            <p th:if="${#fields.hasErrors('nome')}" th:errorclass="text-danger"
-                                th:errors="${usuarioDTO.nome}"></p>
-                        </div>
-                    </div>
-
-                    <div class="row mb-3">
-                        <label class="col-sm-4 col-form-label">Email</label>
-                        <div class="col-sm-8">
-                            <input class="form-control" th:field="${usuarioDTO.email}">
-                            <p th:if="${#fields.hasErrors('email')}" th:errorclass="text-danger"
-                                th:errors="${usuarioDTO.email}"></p>
-                        </div>
-                    </div>
-
-                    <div class="row mb-3">
-                        <label class="col-sm-4 col-form-label">CPF</label>
-                        <div class="col-sm-8">
-                            <input class="form-control" th:field="${usuarioDTO.cpf}">
-                            <p th:if="${#fields.hasErrors('cpf')}" th:errorclass="text-danger"
-                                th:errors="${usuarioDTO.cpf}"></p>
-                        </div>
-                    </div>
-
-                    <div class="row mb-3">
-                        <label class="col-sm-4 col-form-label">Grupo</label>
-                        <div class="col-sm-8">
-                            <select class="form-select" th:field="${usuarioDTO.grupo}">
-                                <option value="Administrador">Administrador</option>
-                                <option value="Estoquista">Estoquista</option>
-                            </select>
-                            <p th:if="${#fields.hasErrors('grupo')}" th:errorclass="text-danger"
-                                th:errors="${usuarioDTO.grupo}"></p>
-                        </div>
-                    </div>
-
-                    <!-- Senha -->
-                    <div class="row mb-3">
-                        <label class="col-sm-4 col-form-label">Senha</label>
-                        <div class="col-sm-8">
-                            <input type="password" th:field="${usuarioDTO.senha}" class="form-control">
-                            <p th:if="${#fields.hasErrors('senha')}" th:errorclass="text-danger"
-                                th:errors="${usuarioDTO.senha}"></p>
-                        </div>
-                    </div>
-
-                    <!-- Confirmar senha -->
-                    <div class="row mb-3">
-                        <label class="col-sm-4 col-form-label">Confirmar senha</label>
-                        <div class="col-sm-8">
-                            <input type="password" id="confirmPassword" class="form-control">
-                        </div>
-                    </div>
-
-                    <div class="row">
-                        <div class="offset-sm-4 col-sm-4 d-grid">
-                            <button type="submit" class="btn btn-primary">Salvar</button>
-                        </div>
-                        <div class="col-sm-4 d-grid">
-                            <a class="btn btn-secondary" href="/usuarios" role="button">Voltar</a>
-                        </div>
-                    </div>
-
-                </form>
-            </div>
-        </div>
-    </div>
-
-    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
-        integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
-        crossorigin="anonymous"></script>
-</body>
-
-</html>
