+++ iOutra_PI4/iOutra/src/main/resources/application.properties
@@ -1,0 +1,22 @@
+spring.application.name=iOutra
+
+# Configuracoes do Banco de dados
+spring.sql.init.platform=h2
+spring.datasource.driver-class-name=org.h2.Driver
+# No Windows, cria o arquivo C:\Users\<NOME_USUARIO>\filmebd-2024-1b.mv
+# Para reiniciar o banco, apagar o arquivo e reiniciar a aplica\u00e7\u00e3o Spring boot
+spring.datasource.url=jdbc:h2:file:~/ioutra-bd
+spring.datasource.username=sa
+spring.datasource.password=
+
+spring.jpa.show-sql=true
+spring.jpa.hibernate.ddl-auto=update
+#spring.jpa.open-in-view=false
+
+# Configuracoes especificas do banco H2
+spring.h2.console.enabled=true
+spring.h2.console.settings.web-allow-others=true
+
+# JACKSON 2 - FORMATAR JSON NA RESPOSTA
+spring.jackson.serialization.INDENT_OUTPUT=true
+
