+++ iOutra_PI4/iOutra/src/main/resources/templates/usuarios/index.html
@@ -1,0 +1,67 @@
+<!DOCTYPE html>
+<html lang="pt-br">
+
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>iOutra - Usuários</title>
+
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/css/bootstrap.min.css" rel="stylesheet"
+        integrity="sha384-QWTKZyjpPEjISv5WaRU9OFeRpok6YctnYmDr5pNlyT2bRjXh0JMhjY6hW+ALEwIH" crossorigin="anonymous">
+</head>
+
+<body>
+    <div class="container">
+        <h1 class="text-center my-4">Lista de Usuários</h1>
+        <div class="container text-center bg-body-secondary my-4">
+
+            <a class="btn btn-primary" href="/usuarios/cadastrarUsuario">Cadastrar Usuário</a>
+            <a class="btn btn-secondary" href="/backoffice">Voltar</a>
+        </div>
+
+        <table class="table ">
+            <thead class="table-dark">
+                <tr>
+                    <th>ID</th>
+                    <th>Nome</th>
+                    <th>CPF</th>
+                    <th>Email</th>
+                    <th>Grupo</th>
+                    <th>Status</th>
+                    <th>Ação</th>
+                </tr>
+            </thead>
+            <tbody class="table-group-divider">
+                <tr th:each="usuario : ${usuarios}">
+                    <td th:text="${usuario.idUsuario}"></td>
+                    <td th:text="${usuario.nome}"></td>
+                    <td th:text="${usuario.cpf}"></td>
+                    <td th:text="${usuario.email}"></td>
+                    <td th:text="${usuario.grupo}"></td>
+                    <td th:text="${usuario.isActive} ? 'Ativo' : 'Inativo'"></td>
+
+                    <td style="white-space: nowrap">
+                        <a class="btn btn-dark btn-sm"
+                            th:href="@{/usuarios/editarUsuario(idUsuario=${usuario.idUsuario})}">Editar</a>
+
+                        <form th:action="@{/usuarios/alterarStatus}" method="post" style="display:inline;">
+                            <input type="hidden" name="id" th:value="${usuario.idUsuario}" />
+                            <button type="submit" class="btn btn-secondary btn-sm"
+                                th:text="${usuario.isActive} ? 'Inativar' : 'Ativar'"
+                                onclick="return confirm('Deseja alterar o status deste usuário?')"></button>
+                        </form>
+
+                    </td>
+
+
+                </tr>
+            </tbody>
+        </table>
+    </div>
+
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.3/dist/js/bootstrap.bundle.min.js"
+        integrity="sha384-YvpcrYf0tY3lHB60NNkmXc5s9fDVZLESaAA55NDzOxhy9GkcIdslK1eN7N6jIeHz"
+        crossorigin="anonymous"></script>
+</body>
+
+</html>
